Separação dos tokens e modificação do função lertokes
Criação de arquivo tokens.py
Integração dos arquivos com o tokens.py
